Fixed #46 and made more responsive entryView
